Implement second half of Rewrite
This completes the Rewrite function.  The change to AnyExp is a bit
of a kludge, but it can do for now.  I am not sure what use case a
matrix wildcard copy would have at the moment.

It would also be possible to construct the new matrix expression using
reflection, but that would be both slower and more complicated.